linux/signal.h: Ignore SIGINFO by default in new tasks
This matches the behaviour of other Unix-like systems that
have SIGINFO and causes less harm to processes that do not
handle this signal, making the keyboard status character
non-fatal for them.

This is implemented with the assumption that SIGINFO is defined
to be equivalent to SIGPWR; still, there is no reason for PWR to
result in termination of the signal recipient anyway — it does not
indicate there is a fatal problem with the recipient's execution
context (like e.g. FPE/ILL do), and we have TERM/KILL for explicit
termination requests.

Signed-off-by: Arseny Maslennikov <ar@cs.msu.ru>